# docs/swagger.yaml
basePath: /
consumes:
- application/x-www-form-urlencoded
definitions:
...
  Board:
    properties:
      board_member_count:
        format: int64
        type: integer
        x-go-name: MemberCount
      gid:
        type: string
        x-go-name: Gid
      id:
        format: int64
        type: integer
        x-go-name: Id
      title:
        type: string
        x-go-name: Title
    type: object
    x-go-package: organizr/server/types
  Error:
    properties:
      code:
        type: string
        x-go-name: Code
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: organizr/server/types
...
host: localhost:1323
info:
  description: API Documentation.
  title: Organizr
  version: 1.0.0
paths:
  /api/board:
    get:
      description: Retrieves all boards
      operationId: board-retrieve-all
      responses:
        "200":
          $ref: '#/responses/multi-board-response'
        "400":
          $ref: '#/responses/error-response'
      security:
      - Bearer:
        - '[]'
      tags:
      - board
    post:
      description: Creates a board
      operationId: board-create
      responses:
        "200":
          $ref: '#/responses/single-board-response'
        "400":
          $ref: '#/responses/error-response'
      security:
      - Bearer:
        - '[]'
      tags:
      - board
  /api/board/{Board_GID}:
    delete:
      description: Deletes board by UUID
      operationId: board-delete
      parameters:
      - description: UUID of board
        in: path
        name: Board_GID
        required: true
        type: string
      responses:
        "200":
          description: ""
        "400":
          $ref: '#/responses/error-response'
      security:
      - Bearer:
        - '[]'
      tags:
      - board
    get:
      description: Retrieves board by UUID
      operationId: board-retrieve-one
      parameters:
      - description: UUID of board
        in: path
        name: Board_GID
        required: true
        type: string
      responses:
        "200":
          $ref: '#/responses/single-board-response'
        "400":
          $ref: '#/responses/error-response'
      security:
      - Bearer:
        - '[]'
      tags:
      - board
    put:
      description: Updates board by UUID
      operationId: board-update
      parameters:
      - description: UUID of board
        in: path
        name: Board_GID
        required: true
        type: string
      - description: Title of board
        in: body
        name: Title
        schema:
          type: string
      responses:
        "200":
          $ref: '#/responses/single-board-response'
        "400":
          $ref: '#/responses/error-response'
      security:
      - Bearer:
        - '[]'
      tags:
      - board
...
produces:
- application/json
responses:
  error-response:
    description: ""
    schema:
      $ref: '#/definitions/Error'
...
  multi-board-response:
    description: ""
    schema:
      items:
        $ref: '#/definitions/Board'
      type: array
...
  single-board-response:
    description: ""
    schema:
      $ref: '#/definitions/Board'
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"